+++ app/src/main/res/layout/activity_menu_info.xml
@@ -5,8 +5,8 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".menuinfo.MenuInfoActivity">
-  
+    tools:context=".menu.menuInfo.MenuInfoActivity">
+
     <FrameLayout
         android:id="@+id/menu_info_frm"
         android:layout_width="match_parent"
